A RouteWrapperBuilder to set the default status
A route can be annotated with @Status(HTTPStatus...) to set the default
response status of the route.  This is useful for exception routes, where
a 404 or 500 response needs to be sent.  An example route would be:

  @Any("/**")
  @Catch(BalsaNotFound.class)
  @Status(HTTPStatus.NotFound)
  public void notFound()
  {
    // Encode the 404 view
    encode("notfound");
  }

Note, when using the ResponseBuilders @JSON and @XML using @Status is
pointless as those ResponseBuilders set the status.