Support data bindings on includes
When including a view data bindings can be setup, this allows for an
included view to make use of variables under a common name and allows
for the value of these variables to be controlled by the consumer of
the include.

Data bindings are specified with 'data-' attributes, where the
identifier following the 'data-' prefix is the variable name which
will be set.

For example, we could include a view:

<include data-message="Hello World" view="test/hello"/>

In the view 'test/hello', we can then do:

<p>#{message}</p>